Add LZCompression utility
Compression is useful for sending data to and from pucks, as achievable
BLE link throughput is quite low. The perfect compression algorithm for
this case would have good compression and a small memory footprint, as
RAM is on the premium on the embedded side of things.  LZ compression
scores well on both of these things (no working memory required for
decompression!), which is why it is chosen over, say, Deflate, which
exists already as a part of the core java libraries.

The LZCompression implementation introduced in this commit is based off
of the LZ compression implementation from the Basic Compression
Library ( http://bcl.comli.eu/ ).