Add custom Bluetooth GattManager
The Bluetooth LE stack on Android is seriously flaky, so a custom
BLE manager is needed for method wrapping and error workarounds. The
GattManager maintains a queue of GattOperations to be performed, and
executes them when it feels it is safe to do so.

In order to perform Gatt operations, users (i.e. developers) should
simply queue the operations they want to execute, and let the manager
do its job.