Make GattManager more robust
This commit improves GattManager robustness by operation execution in
a convenient wrapper, using now-explicit information about whether or
not it should expect an operation to return via a callback when it is
complete. Operation timeout is also codified explicitly in each
operation, but the GattManager does currently not honor it.